add facebook login buttons and login flow
You can log in and it'll show your picture and name at the top of the
feed. Having some issues with redirects after login. Once you log in,
there's a button that should take you back to the feed but it doesn't do
anything. And when you go back to the pledges page, it still shows that
button